+++ test/src/test/kotlin/graphql/nadel/tests/next/fixtures/basic/BasicObjectSchemaTestData.kt
@@ -1,5 +1,5 @@
 // @formatter:off
-package graphql.nadel.tests.next.fixtures
+package graphql.nadel.tests.next.fixtures.basic
 
 import graphql.nadel.tests.next.ExpectedNadelResponse
 import graphql.nadel.tests.next.ExpectedServiceCall
