+++ test/src/test/kotlin/graphql/nadel/tests/next/fixtures/basic/EchoTest.kt
@@ -1,4 +1,4 @@
-package graphql.nadel.tests.next.fixtures
+package graphql.nadel.tests.next.fixtures.basic
 
 import graphql.nadel.tests.next.NadelIntegrationTest
