+++ test/src/test/kotlin/graphql/nadel/tests/next/fixtures/hydration/HydrationTest.kt
@@ -1,4 +1,4 @@
-package graphql.nadel.tests.next.fixtures
+package graphql.nadel.tests.next.fixtures.hydration
 
 import graphql.nadel.tests.next.NadelIntegrationTest
