+++ test/src/test/kotlin/graphql/nadel/tests/next/fixtures/hydration/defer/HydrationDeferInListIsDisabledForParentIssueInRelatedIssuesTestData.kt
@@ -1,5 +1,5 @@
 // @formatter:off
-package graphql.nadel.tests.next.fixtures
+package graphql.nadel.tests.next.fixtures.hydration.defer
 
 import graphql.nadel.tests.next.ExpectedNadelResponse
 import graphql.nadel.tests.next.ExpectedServiceCall
@@ -15,7 +15,7 @@ import kotlin.collections.listOf
  * Refer to [graphql.nadel.tests.next.CaptureTestData]
  */
 @Suppress("unused")
-public class DeferHydrationInListIsDisabledForParentIssueInRelatedIssuesTestData : TestData() {
+public class HydrationDeferInListIsDisabledForParentIssueInRelatedIssuesTestData : TestData() {
     override val calls: List<ExpectedServiceCall> = listOf(
             ExpectedServiceCall(
                 service = "issues",
